Create search bar and style it
In the HTML the input was wrapped in the FORM element, in which the
action sends the information to the Google engine and the method is
get.

The button to submit the information has an image to give a better
appearance to the input. On the other hand, the form container is used
as the input visualization.