Create second level of the header
The HTML was changed in some sections to accomplish
this. In the HTML file the container for this level was inside the div
for the first level, thus it was taken out of there and placed in a
corresponding div so the container could exapand from the second
line row to the third.

In the CSS file the layout of this level was styled and arranged.