Create responsive rows and columns for cards
The cards contained in the area where the user can have his cards with
custom information now are able to create columns and rows according
to the space available. This was made by using the repeat() function
in conjunction with auto-fit and the minmax() function when creating
the rows and columns.

Likewise, each card was given a white background and border-radius.