Set height for cards and add container for bottom options
The height of the cards was set to prevent implicit elements from being
shrinked. On the other hand, within the cards was missing a container
for the options available, that is, if the user wants to share the card,
save it as a favorite, etc. Thus, in the HTML file those divs were
added.