Add ellipsis to prevent overflow across vertical axis
This was made possible with the help of -webkit-box in the cards, which
allowed to specify the number of lines that the text is allowed to
span, but once that limit is surpased it ends with three dots.

The titles of the cards were given a bigger font as well as the titles
of these section, that is: Your projects, Announcements, and Trendings.